Split helpers and re-wrote exceptions file
