+++ praw/models/reddit/live.py
@@ -77,8 +77,7 @@ class LiveContributorRelationship:
     ):
         """Invite a redditor to be a contributor of the live thread.
 
-        Raise :class:`praw.exceptions.RedditAPIException` if the invitation
-        already exists.
+        Raise :class:`.RedditAPIException` if the invitation already exists.
 
         :param redditor: A redditor name (e.g., ``'spez'``) or
             :class:`~.Redditor` instance.
